requirements.txt
# Core AI/ML Libraries
langgraph==0.2.16
langchain==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-anthropic==0.2.0

# Neural Networks
torch==2.1.2
brian2==2.6.0
numpy==1.24.3
scipy==1.11.4

# Data Processing
pandas==2.1.4
scikit-learn==1.3.2

# MCP (Model Context Protocol)
mcp==0.9.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# Web Framework (for real-time data API)
flask==3.0.0
flask-cors==4.0.0
websockets==12.0

# Vector Database
chromadb==0.4.22

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Visualization (for demos)
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# Database
sqlalchemy==2.0.25
